[#5] 학번이 아니라 id로 db조회, JPA로 레포코드수정

+++ satto/src/main/java/insung/satto/domain/user/entity/User.java
@@ -15,9 +15,8 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
-//    @Column(name = "studentId", length = 20)
     private String studentId;
     private String password;
     private String username;
